feat: 서재 페이지 이동 자체는 가능하게 연동
