+++ frontend/tsconfig.json
@@ -26,6 +26,6 @@
       "@interfaces": ["./interfaces"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "components/study/BookListTab"],
   "exclude": ["node_modules"]
 }
